Added support for OpenSSL 1.1.
This commit adds compatibility with OpenSSL 1.1. In h235 implementation,
CTS crypto engine has been reworked as a helper class which stores the
necessary cipher members that are not available publicly in OpenSSL 1.1
locally. In other places, added OpenSSL 1.1 API equivalents to use with
older OpenSSL versions.

Updated thread cleanup on older OpenSSL versions to use ERR_remove_thread_state.
On OpenSSL 1.1 no cleanup is needed as the library does it itself.

In TLS context initialization also disable TLS compression as it is vulnerable
to CRIME.