Add support for OpenSSL 1.1.0 to h235.
This commit adds compatibility with OpenSSL 1.1.0 to h235 implementation.
The CTS crypto functions were reworked as a helper class which stores the
necessary cipher members that are not available publicly in OpenSSL 1.1.0
locally.

Unlike 9a65ea28ba50ce6fe8239fa7e2c69c5086be7aba, this commit also adds
a proper non-CTS DecryptUpdate implementation, which is needed to keep the
internal buffer in the helper updated until it is used in DecryptFinalRelaxed.
This implementation, which also includes EncryptUpdate, is based on the
code from OpenSSL.