Avoid out-of-bounds access
According to encoder logic speech_mode is in the range 0..8.
D_IF_homing_frame_test_first relies on that; i.e. if speech_mode is out of mentioned
range, then out-of-bounds access happens.

This PR adds a check that speech_mode is valid right after it is parsed;
in case it is invalid, current frame is marked as invalid (RX_SPEECH_LOST);
as a result usual logic for decoding recovery is applied.